refactor: Improve Todos component with remove todo functionality
The Todos component has been refactored to include the ability to remove individual todos. The `Todos` component now accepts an `onRemoveTodo` prop, which is a function that takes a todo ID as an argument. This allows users to remove todos by clicking on them. This change enhances the functionality and user experience of the Todos component.